+++ lib/logstash/inputs/eventlog.rb
@@ -4,6 +4,8 @@ require "logstash/namespace"
 require "logstash/timestamp"
 require "win32/eventlog"
 require "stud/interval"
+require "logstash/util/charset"
+
 
 # This input will pull events from a http://msdn.microsoft.com/en-us/library/windows/desktop/bb309026%28v=vs.85%29.aspx[Windows Event Log].
 # Note that Windows Event Logs are stored on disk in a binary format and are only accessible from the Win32 API.
@@ -32,6 +34,10 @@ class LogStash::Inputs::EventLog < LogStash::Inputs::Base
   # How frequently should tail check for new event logs in ms (default: 1 second)
   config :interval, :validate => :number, :default => 1000
 
+  # Event Log string encoding (default: UTF-16LE), however your system might be using another encoding, if you
+  # are seeing strange characters, inspect this variable.
+  config :charset, :validate => :string, :default => "UTF-16LE"
+
   public
   def register
 
@@ -47,6 +53,7 @@ class LogStash::Inputs::EventLog < LogStash::Inputs::Base
       end
       raise
     end
+    @converter = LogStash::Util::Charset.new(Encoding.find(@charset))
   end # def register
 
   public
@@ -73,7 +80,7 @@ class LogStash::Inputs::EventLog < LogStash::Inputs::Base
   private
   def process(log)
 
-    LogStash::Event.new(
+    attrs = {
       "host"             => @hostname,
       "Logfile"          => @logfile,
       "message"          => log["description"].strip,
@@ -88,8 +95,25 @@ class LogStash::Inputs::EventLog < LogStash::Inputs::Base
       "Type"             => log["event_type"],
       "User"             => log["user"],
       "InsertionStrings" => log["string_inserts"]
-    )
+    }
+
+    attrs.each do |k,v|
+      next if ["host", "Logfile"].include?(k)
+      attrs[k] = convert(v)
+    end
+
+    LogStash::Event.new(attrs)
   end # def run
+
+  def convert(field)
+    if field.is_a?(String)
+      @converter.convert(field)
+    elsif field.is_a?(Array)
+      field.map { |v| @converter.convert(v) }
+    else
+      field
+    end
+  end
 
 end # class LogStash::Inputs::EventLog